+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Added Swagger documentation on the [API](https://llm-guard.com/usage/api/) documentation page
+- Added `fail_fast` flag to stop the execution after the first failure
 
 ### Removed
 -